Fix babel-plugin-transform-es2015-modules-systemjs tests
