+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-default-5/expected.js
@@ -1,0 +1,10 @@
+System.register([], function (_export, _context) {
+  "use strict";
+
+  _export("default", function () {});
+
+  return {
+    setters: [],
+    execute: function () {}
+  };
+});
