+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-default-7/actual.js
@@ -1,0 +1,1 @@
+export default function foo () {}
