+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-default-7/expected.js
@@ -1,0 +1,12 @@
+System.register([], function (_export, _context) {
+  "use strict";
+
+  function foo() {}
+
+  _export("default", foo);
+
+  return {
+    setters: [],
+    execute: function () {}
+  };
+});
