+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-default-9/actual.js
@@ -1,0 +1,1 @@
+export default (function(){return "foo"})();
