+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-default-9/expected.js
@@ -1,0 +1,12 @@
+System.register([], function (_export, _context) {
+  "use strict";
+
+  return {
+    setters: [],
+    execute: function () {
+      _export("default", function () {
+        return "foo";
+      }());
+    }
+  };
+});
