+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-default/expected.js
@@ -1,0 +1,10 @@
+System.register([], function (_export, _context) {
+  "use strict";
+
+  return {
+    setters: [],
+    execute: function () {
+      _export("default", 42);
+    }
+  };
+});
