+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-from-2/actual.js
@@ -1,0 +1,1 @@
+export {foo} from "foo";
