+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-from-3/expected.js
@@ -1,0 +1,14 @@
+System.register(["foo"], function (_export, _context) {
+  "use strict";
+
+  return {
+    setters: [function (_foo) {
+      var _exportObj = {};
+      _exportObj.foo = _foo.foo;
+      _exportObj.bar = _foo.bar;
+
+      _export(_exportObj);
+    }],
+    execute: function () {}
+  };
+});
