+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-from-5/expected.js
@@ -1,0 +1,13 @@
+System.register(["foo"], function (_export, _context) {
+  "use strict";
+
+  return {
+    setters: [function (_foo) {
+      var _exportObj = {};
+      _exportObj.default = _foo.foo;
+
+      _export(_exportObj);
+    }],
+    execute: function () {}
+  };
+});
