+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-from/actual.js
@@ -1,0 +1,1 @@
+export * from "foo";
