+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-from/expected.js
@@ -9,14 +9,6 @@ System.register(["foo"], function (_export, _context) {
         if (_key !== "default" && _key !== "__esModule") _exportObj[_key] = _foo[_key];
       }
 
-      _exportObj.foo = _foo.foo;
-      _exportObj.foo = _foo.foo;
-      _exportObj.bar = _foo.bar;
-      _exportObj.bar = _foo.foo;
-      _exportObj.default = _foo.foo;
-      _exportObj.default = _foo.foo;
-      _exportObj.bar = _foo.bar;
-
       _export(_exportObj);
     }],
     execute: function () {}
