+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-named-2/actual.js
@@ -1,0 +1,1 @@
+export {foo, bar};
