+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-named-2/expected.js
@@ -1,0 +1,12 @@
+System.register([], function (_export, _context) {
+  "use strict";
+
+  return {
+    setters: [],
+    execute: function () {
+      _export("foo", foo);
+
+      _export("bar", bar);
+    }
+  };
+});
