+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-named-6/actual.js
@@ -1,0 +1,1 @@
+export function foo() {}
