+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-named-6/expected.js
@@ -1,0 +1,12 @@
+System.register([], function (_export, _context) {
+  "use strict";
+
+  function foo() {}
+
+  _export("foo", foo);
+
+  return {
+    setters: [],
+    execute: function () {}
+  };
+});
