+++ packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/export-named-7/expected.js
@@ -1,0 +1,12 @@
+System.register([], function (_export, _context) {
+  "use strict";
+
+  function foo2(bar) {}
+
+  _export("foo2", foo2);
+
+  return {
+    setters: [],
+    execute: function () {}
+  };
+});
