--- packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/exports-default/actual.js
@@ -1,9 +1,0 @@
-export default 42;
-export default {};
-export default [];
-export default foo;
-export default function () {}
-export default class {}
-export default function foo () {}
-export default class Foo {}
-export default (function(){return "foo"})();
