--- packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/exports-from/actual.js
@@ -1,6 +1,0 @@
-export * from "foo";
-export {foo} from "foo";
-export {foo, bar} from "foo";
-export {foo as bar} from "foo";
-export {foo as default} from "foo";
-export {foo as default, bar} from "foo";
