--- packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/exports-named/actual.js
@@ -1,7 +1,0 @@
-export {foo};
-export {foo, bar};
-export {foo as bar};
-export {foo as default};
-export {foo as default, bar};
-export function foo() {}
-export function foo2(bar) {}
