--- packages/babel-plugin-transform-es2015-modules-systemjs/test/fixtures/systemjs/exports-named/expected.js
@@ -1,30 +1,0 @@
-System.register([], function (_export, _context) {
-  "use strict";
-
-  function foo() {}
-
-  _export("foo", foo);
-
-  function foo2(bar) {}
-
-  _export("foo2", foo2);
-
-  return {
-    setters: [],
-    execute: function () {
-      _export("foo", foo);
-
-      _export("foo", foo);
-
-      _export("bar", bar);
-
-      _export("bar", foo);
-
-      _export("default", foo);
-
-      _export("default", foo);
-
-      _export("bar", bar);
-    }
-  };
-});
